HAPP 636 - added option to logout from profile - removed option to logout from setting - fixed delete all record option by logging out from BCSC.
