+++ app/src/main/res/drawable/ic_log_out.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="21dp"
+    android:height="19dp"
+    android:viewportWidth="21"
+    android:viewportHeight="19">
+  <path
+      android:pathData="M7.8779,17.6406C7.8779,18.1861 7.4389,18.625 6.8931,18.625H3.9389C1.7635,18.625 0,16.8621 0,14.6875V4.1875C0,2.0129 1.7635,0.25 3.9389,0.25H6.8931C7.4389,0.25 7.8779,0.6905 7.8779,1.2344C7.8779,1.7778 7.4389,2.2188 6.8931,2.2188H3.9389C2.8557,2.2188 1.9695,3.1047 1.9695,4.1875V14.6875C1.9695,15.7703 2.8557,16.6562 3.9389,16.6562H6.8931C7.4389,16.6562 7.8779,17.0951 7.8779,17.6406ZM20.741,8.7648L15.526,3.1867C15.1529,2.7894 14.5298,2.7716 14.1346,3.1444C13.7385,3.5161 13.7194,4.1391 14.0923,4.5353L17.7416,8.4531H7.5127C7.0039,8.4531 6.5649,8.8961 6.5649,9.4375C6.5649,9.9789 7.006,10.4219 7.5127,10.4219H17.7047L14.0152,14.3413C13.6421,14.7374 13.6614,15.3602 14.0575,15.7322C14.2869,15.9098 14.529,16 14.7341,16C14.9957,16 15.2572,15.8962 15.4509,15.6898L20.6659,10.1117C21.098,9.7328 21.098,9.1422 20.741,8.7648Z"
+      android:fillColor="#D72E3D"/>
+</vector>
