+++ app/src/main/res/values/colors.xml
@@ -16,6 +16,7 @@
     <color name="shadow_grey">#CFCFCF</color>
     <color name="light_grey">#F7F7F7</color>
     <color name="error">#D8292F</color>
+    <color name="red">#D72E3D</color>
 
     <!--Covid Test Result colors-->
     <color name="covid_test_blue">#D2E2EF</color>
